refactor: UserState replace for selfUser and some code cleanups
