+++ src/script/page/AppMain.tsx
@@ -236,6 +236,7 @@ const AppMain: FC<AppMainProps> = ({
                 actionsViewModel={mainView.actions}
                 isFederated={mainView.isFederated}
                 teamState={teamState}
+                selfUser={selfUser}
                 userState={userState}
               />
             )}
