+++ src/script/page/LeftSidebar/LeftSidebar.tsx
@@ -56,14 +56,10 @@ const LeftSidebar: React.FC<LeftSidebarProps> = ({listViewModel, selfUser, isAct
   const switchList = (list: ListState) => listViewModel.switchList(list);
 
   const goHome = () =>
-    selfUser.isTemporaryGuest()
-      ? listViewModel.switchList(ListState.TEMPORARY_GUEST)
-      : listViewModel.switchList(ListState.CONVERSATIONS);
+    selfUser.isTemporaryGuest() ? switchList(ListState.TEMPORARY_GUEST) : switchList(ListState.CONVERSATIONS);
 
   useEffect(() => {
-    amplify.subscribe(WebAppEvents.SHORTCUT.START, () => {
-      listViewModel.switchList(ListState.START_UI);
-    });
+    amplify.subscribe(WebAppEvents.SHORTCUT.START, () => switchList(ListState.START_UI));
   }, []);
 
   return (
@@ -118,6 +114,7 @@ const LeftSidebar: React.FC<LeftSidebarProps> = ({listViewModel, selfUser, isAct
                 mainViewModel={listViewModel.mainViewModel}
                 userRepository={repositories.user}
                 isFederated={listViewModel.isFederated}
+                selfUser={selfUser}
               />
             )}
           </>
