+++ src/script/page/MainContent/panels/Collection/Collection.tsx
@@ -17,7 +17,7 @@
  *
  */
 
-import React, {useEffect, useState} from 'react';
+import {useEffect, useState} from 'react';
 
 import {amplify} from 'amplify';
 
@@ -29,7 +29,6 @@ import {ConversationRepository} from 'src/script/conversation/ConversationReposi
 import {ContentMessage} from 'src/script/entity/message/ContentMessage';
 import {generateConversationUrl} from 'src/script/router/routeGenerator';
 import {createNavigate} from 'src/script/router/routerBindings';
-import {UserState} from 'src/script/user/UserState';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 
@@ -41,6 +40,7 @@ import {Category, isOfCategory} from './utils';
 import {AssetRepository} from '../../../../assets/AssetRepository';
 import {MessageRepository} from '../../../../conversation/MessageRepository';
 import {Conversation} from '../../../../entity/Conversation';
+import {User} from '../../../../entity/User';
 import {MessageCategory} from '../../../../message/MessageCategory';
 
 interface CollectionDetailsProps {
@@ -48,7 +48,7 @@ interface CollectionDetailsProps {
   conversationRepository: ConversationRepository;
   assetRepository: AssetRepository;
   messageRepository: MessageRepository;
-  readonly userState: UserState;
+  selfUser: User;
 }
 
 type Categories = Record<Category, ContentMessage[]>;
@@ -80,18 +80,17 @@ function splitIntoCategories(messages: ContentMessage[]): Categories {
   );
 }
 
-const Collection: React.FC<CollectionDetailsProps> = ({
+const Collection = ({
   conversation,
   conversationRepository,
   assetRepository,
   messageRepository,
-  userState,
-}) => {
+  selfUser,
+}: CollectionDetailsProps) => {
   const [searchTerm, setSearchTerm] = useState('');
   const {display_name} = useKoSubscribableChildren(conversation, ['display_name']);
   const [messages, setMessages] = useState<ContentMessage[]>([]);
   const [detailCategory, setDetailCategory] = useState<Category | undefined>(undefined);
-  const {self: selfUser} = useKoSubscribableChildren(userState, ['self']);
 
   useEffect(() => {
     conversationRepository
